Organize pages into folders with pastel gradient
